move render functions to separate files
