+++ frontend/src/Layout.js
@@ -1,0 +1,9 @@
+const Layout = () => {
+  return (
+    <div>
+      <h1>FileChest</h1>
+    </div>
+  );
+};
+
+export default Layout
